refactor: use dedicated vector index control client and grpc manager
Previously we comingled the vector index control client with the cache
client code. This is because the vector index commands reside on the
cache control service.

Unfortunately, because the vector index service and cache service have
different configuration needs, and the control client takes as input a
cache configuration object, we need to separate these out.

The upshot is this adds clarity to the project. The vector index grpc
related clients can be found in: _vector_index_grpc_manager,
_vector_index_data_client, and _vector_index_control_client. The files
contain only those commands relevant to the vector index service.
