+++ pyproject.toml
@@ -105,7 +105,9 @@ module = [
   "momento.internal.synchronous._scs_control_client",
   "momento.internal.synchronous._scs_data_client",
   "momento.internal.synchronous._scs_grpc_manager",
+  "momento.internal.synchronous._vector_index_control_client",
   "momento.internal.synchronous._vector_index_data_client",
+  "momento.internal.synchronous._vector_index_grpc_manager",
   "momento.internal.synchronous._add_header_client_interceptor",
   "momento.internal.synchronous._retry_interceptor",
   "momento.internal.common._data_client_ops",
@@ -115,7 +117,9 @@ module = [
   "momento.internal.aio._scs_control_client",
   "momento.internal.aio._scs_data_client",
   "momento.internal.aio._scs_grpc_manager",
+  "momento.internal.aio._vector_index_control_client",
   "momento.internal.aio._vector_index_data_client",
+  "momento.internal.aio._vector_index_grpc_manager",
   "momento.internal.aio._utilities",
   "momento.responses.control.signing_key.*",
 ]
