+++ src/momento/internal/aio/_scs_control_client.py
@@ -8,12 +8,7 @@ from momento import logs
 from momento.auth import CredentialProvider
 from momento.config import Configuration
 from momento.errors import convert_error
-from momento.internal._utilities import (
-    _validate_cache_name,
-    _validate_index_name,
-    _validate_num_dimensions,
-    _validate_ttl,
-)
+from momento.internal._utilities import _validate_cache_name, _validate_ttl
 from momento.internal.aio._scs_grpc_manager import _ControlGrpcManager
 from momento.responses import (
     CacheFlush,
@@ -30,14 +25,6 @@ from momento.responses import (
     ListSigningKeysResponse,
     RevokeSigningKey,
     RevokeSigningKeyResponse,
-)
-from momento.responses.vector_index import (
-    CreateIndex,
-    CreateIndexResponse,
-    DeleteIndex,
-    DeleteIndexResponse,
-    ListIndexes,
-    ListIndexesResponse,
 )
 
 _DEADLINE_SECONDS = 60.0  # 1 minute
@@ -70,20 +57,6 @@ class _ScsControlClient:
             return CreateCache.Error(convert_error(e))
         return CreateCache.Success()
 
-    async def create_index(self, index_name: str, num_dimensions: int) -> CreateIndexResponse:
-        try:
-            self._logger.info(f"Creating index with name: {index_name}")
-            _validate_index_name(index_name)
-            _validate_num_dimensions(num_dimensions)
-            request = ctrl_pb._CreateIndexRequest(index_name=index_name, num_dimensions=num_dimensions)
-            await self._build_stub().CreateIndex(request, timeout=_DEADLINE_SECONDS)
-        except Exception as e:
-            self._logger.debug("Failed to create index: %s with exception: %s", index_name, e)
-            if isinstance(e, grpc.RpcError) and e.code() == grpc.StatusCode.ALREADY_EXISTS:
-                return CreateIndex.IndexAlreadyExists()
-            return CreateIndex.Error(convert_error(e))
-        return CreateIndex.Success()
-
     async def delete_cache(self, cache_name: str) -> DeleteCacheResponse:
         try:
             self._logger.info(f"Deleting cache with name: {cache_name}")
@@ -95,17 +68,6 @@ class _ScsControlClient:
             return DeleteCache.Error(convert_error(e))
         return DeleteCache.Success()
 
-    async def delete_index(self, index_name: str) -> DeleteIndexResponse:
-        try:
-            self._logger.info(f"Deleting index with name: {index_name}")
-            _validate_index_name(index_name)
-            request = ctrl_pb._DeleteIndexRequest(index_name=index_name)
-            await self._build_stub().DeleteIndex(request, timeout=_DEADLINE_SECONDS)
-        except Exception as e:
-            self._logger.debug("Failed to delete index: %s with exception: %s", index_name, e)
-            return DeleteIndex.Error(convert_error(e))
-        return DeleteIndex.Success()
-
     async def list_caches(self) -> ListCachesResponse:
         try:
             list_caches_request = ctrl_pb._ListCachesRequest(next_token="")
@@ -113,14 +75,6 @@ class _ScsControlClient:
             return ListCaches.Success.from_grpc_response(response)
         except Exception as e:
             return ListCaches.Error(convert_error(e))
-
-    async def list_indexes(self) -> ListIndexesResponse:
-        try:
-            list_indexes_request = ctrl_pb._ListIndexesRequest(next_token="")
-            response = await self._build_stub().ListIndexes(list_indexes_request, timeout=_DEADLINE_SECONDS)
-            return ListIndexes.Success.from_grpc_response(response)
-        except Exception as e:
-            return ListIndexes.Error(convert_error(e))
 
     async def flush(self, cache_name: str) -> CacheFlushResponse:
         try:
